add preliminary 32bit support
use uintptr_t instead of uint64_t
allow multiple %l in printf
select types based on arch in printf
use sv32 virtual addressing on riscv32
link to libgcc on riscv32
use architecture dependent register ASM load/store size
use %ll printf for uint64_t's
use %p printf for addresses + registers
add make target to print all c defines

+++ src/kernel/process.h
@@ -25,15 +25,15 @@ struct Process {
     ProcessState state;
     uint8_t kernel_stack[KERNEL_STACK_SIZE];
     TrapFrame context;
-    uint64_t program_counter;
+    uintptr_t program_counter;
     PageTable page_table;
 };
 typedef struct Process Process;
 
 struct ProcessArguments {
-    uint64_t entry_address;
+    uintptr_t entry_address;
     bool is_user_program;
-    uint64_t user_program_end;
+    uintptr_t user_program_end;
 };
 typedef struct ProcessArguments ProcessArguments;
